backend/requirements.txt
transformers
accelerate
tensorboard
pillow
timm
#eniops
flask >= 3.1.0
flask-restful == 0.3.10
python-dotenv == 1.0.1
flasgger == 0.9.7.1
werkzeug == 3.1.3
flask-cors == 5.0.0
python-dateutil == 2.8.2
